+++ tabletest-junit/src/main/java/io/github/nchaugen/tabletest/junit/FactorySources.java
@@ -22,6 +22,6 @@ import java.lang.annotation.Target;
 
 @Target({ElementType.ANNOTATION_TYPE, ElementType.TYPE})
 @Retention(RetentionPolicy.RUNTIME)
-public @interface TableTestConverters {
+public @interface FactorySources {
     Class<?>[] value();
 }